Moved scraping classes into separate file and added Look Fantastic brand list scraper.
